requirements.txt
blinker==1.9.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.1.0
flask-cors==5.0.1
folium==0.19.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
narwhals==1.31.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
plotly==6.0.1
pycryptodome==3.21.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
setuptools==77.0.3
simplekml==1.3.6
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2025.1.0
